#94 added early detection of bad job files in emails and routing of error back to sender, integration tested it

+++ src/it/java/eu/openanalytics/rsb/RestITCase.java
@@ -103,6 +103,10 @@ public class RestITCase extends AbstractITCase {
             }
         });
 
+        if (testResultFiles == null) {
+            return;
+        }
+
         for (final File testResultFile : testResultFiles) {
             FileUtils.deleteQuietly(testResultFile);
         }
